apps/BACKUP UI 2/requirements.txt
# UI Framework
panel>=1.3.0
param>=2.0.0
bokeh>=3.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# HTTP Client
httpx>=0.25.0

# Authentication & Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Environment & Configuration
python-dotenv>=1.0.0

# Date & Time
python-dateutil>=2.8.0

# JSON & Data
pydantic>=2.0.0

# Utilities
aiofiles>=23.0.0
